added handleHomeRoute and corresponding test :zap: relates #17 #12
